requirements.txt
# GreenSVC-AI 依赖

# Web框架
gradio>=4.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
exifread>=3.0.0

# API客户端
requests>=2.31.0
google-generativeai>=0.8.0  # Gemini API (Stage 1推荐)
openai>=1.0.0  # 用于报告AI分析

# 配置管理
python-dotenv>=1.0.0

# 报告生成
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0